Updated README.
Added outline of program flow, showing the initialization of the
population and its continually evolution until termination.

Wrote stub for arbitrary Algorithm class that demonstrates the creation of
the next population based on the given one. Simply uses the selector to
retrieve parent chromosomes, which after crossing over within another
parent and mutating the result are added as a member of the new
population.