Changed crossover() method.
At least one parent gene must be provided when performing a crossover. The
first parent gene is copied to produce the child gene. Then, each other
parent gene is crossed-over with the child gene using the same method as
in the previous implementation.

Added support for specifying the probability that a crossover occurs.