Created AbstractGeneticAlgorithm class.
Evolves next population using the specified configuration of crossover and
mutation operators. Creates a default population and adds default
chromosomes with genes determined by the crossover and mutator.

Implementation arbitrarily decides to attempt to crossover the parent at
a particular index, and the parent in the index next to it, wrapping back
to the first element when crossing-over the last.