Refactored RouletteWheelSelector class; fixes #4.
Changed select() method to generate an array of probabilities, and then
sorts them. This allows for the nested loops in the previous
implementation to be unwrapped, where the fitnesses/chromosomes arrays
and the probabilities/selected array are iterated through
simulataneously.

A chromosome is selected when the sum of its fitness and all the
chromosomes previous to it is less than the probability.