Changed the signature of crossover() method.
Now takes all parents as a single list since using generics with varargs
is potentially unsafe. The requirement that more than one parent must be
provided can be easily done by checking the size of this list.

Also has the method return a list of genes, rather than a single gene.
This is so that the ``reciprocal'' crossover gene is included as a part
of the next population.

The implementation in NpointCrossover class was changed so that the list
of parents genes was transformed into a map, similar to:
  {gene1: [gene2, ...], gene2: [gene1, ...], ...}
In a similar manner to the previous implementation, a set of points are
generated and a slice of bases are swapped with the other genes.

Since the crossover and mutation operators are applied on genes and not
chromosomes, it possible that one may desire differing operations to be
performed, especially because they may be of different types. The
configuration has been modified in order to support this. It is now
backed by an array of probabilities and a list of operators.