Updated documentation.
Added doc comments for all constructors, methods, and type parameters of
GeneticAlgorithmConfiguration class and its static Builder class.

Changed implementation to create new fields to store the properties of
the configuration, rather than storing the builder instance and retrieve
the properties from it.

This allows for a single builder to create different configurations.