Updated OneMaxExample class.
Changed to use ThresholdTerminationGeneticAlgorithm instance, instead of
FixedTerminationGeneticAlgorithm since the fitness of the optimal
solution is known.

Updated to using new form (specifying `index` parameter) for
GeneticAlgorithmConfiguration.

Changed implementation in AbstractExample class such that the population
is not initialized in the execute() method, but should be done so
beforehand, i.e. in the createPopulation() method.

Suppressed warnings for static and static-access.