Created AbstractCache class.
Backed by a WeakHashMap so that when the instance is no longer used, it
is removed from the cache. This is generally not how a cache works, but
the purpose is to store the fitness values for members of the
population, so once that population evolves those values are usually not
used again.

However, since it is possible to have the same chromosome (although new
instance) exist from one population to the next, the put() method will
actually replace the key stored in the WeakHashMap by removing it and
then add an entry with the specified key and value.

The AbstractCache class has no policy for how to handle a full cache,
hence why it is abstract. In a future revision, an exception will be
throw when trying to add to a full cache.