Fixed bug with NpointCrossover crossover() method.
After implementing the equals() and hashcode() methods for the various
Gene instances, map which stored the relation of each child to the other
parents may contain fewer entries than the number of parents.

This would cause an identical child to be produced from the crossover,
even though the order in which the other parents were crossed over is
different, because putting the second child into the map replaces the
first. Instead, a list children and a list of other parents are used.