Prevented issue with AbstractGeneticAlgorithm class.
Changed implementation to remove assumption that the number of children
returned in the list from the crossover() method is equal to the number
of parents given in the list to the crossover() method. This enables
support for crossover operators such as arithmetic crossover operators,
which perform an operation on many genes and return a single gene.