Cleaned up UniformCrossover class.
Created helper methods to handle the logic of (1) generating an array of
positions that holds what index of the parent is used for the
corresponding child, (2) shuffling the result of (1), and (3) replacing
the value for a gene over all children genes and positions, using each
parent once.

Updated documentation by adding doc comments for all constructors,
methods, and type parameters of UniformCrossover class.