Fixed critical issue with WeakHashMap backing cache.
Retrieve `null` values  due to inconsistencies of multiple calls of the
WeakHashMap methods caused by the garbage collector silently removing
entries. This made it possible for the cache return a `null` value on
the call to get(), even though the call to containsKey() returned
`true`.

One way to fix this would be to only call to get() and upon receiving a
`null` value to recompute the fitness and then make a call to put().
However, because it is possible for this value to not exist on its next
attempt to be retreived due to garbage collection, decided to back the
AbstractCache class with a HashMap instance instead.