Added method to CrossoverOperator interface.
Created getNumChildren(int) method, which returns the number of children
produced by the crossover operation for the given number of parents.

For both NpointCrossover and UniformCrossover classes, this is simply
the number of parents; however, as stated previously
(342902d056ee767cd65575f919823ba7464a865a) it could potentially be
different for arithmetic crossover operators.