Updated specifications for GeneticAlgorithm interface.
Changed evolve() function such that the next population is provided as
empty via a PopulationFactory instance, rather than always using the
DefaultPopulation class.

Modified the apply() functions to replace the selection, crossover, and
mutation operators used by the genetic algorithm instead of adding to a
list of operators that are applied.